+++ src/axom/quest/interface/signed_distance.cpp
@@ -136,21 +136,22 @@ int signed_distance_init(const std::string& file, MPI_Comm comm)
   // STEP 0: read the STL mesh
   int rc = INIT_FAILED;
 
-#if defined(AXOM_USE_UMPIRE_SHARED_MEMORY)
-  if(s_allocator_id == INVALID_ALLOCATOR_ID)
-  {
-    // Make a shared memory allocator and get its id.
-    auto& rm = umpire::ResourceManager::getInstance();
-    auto traits{umpire::get_default_resource_traits("SHARED")};
-    traits.scope = umpire::MemoryResourceTraits::shared_scope::node;
-    auto node_allocator{rm.makeResource("SHARED::node_allocator", traits)};
-    auto signed_distance_allocator{
-      rm.makeAllocator<umpire::strategy::NamedAllocationStrategy>("signed_distance_allocator", node_allocator)};
-    s_allocator_id = signed_distance_allocator.getId();
-  }
-
   if(Parameters.use_shared_memory)
   {
+#if defined(AXOM_USE_UMPIRE_SHARED_MEMORY)
+    if(s_allocator_id == INVALID_ALLOCATOR_ID)
+    {
+      // Make a shared memory allocator if we have not made it before. We'll reuse
+      // the allocator to allocate different buffers (1 at a time).
+      auto& rm = umpire::ResourceManager::getInstance();
+      auto traits{umpire::get_default_resource_traits("SHARED")};
+      traits.scope = umpire::MemoryResourceTraits::shared_scope::node;
+      auto node_allocator{rm.makeResource("SHARED::node_allocator", traits)};
+      auto signed_distance_allocator{
+        rm.makeAllocator<umpire::strategy::NamedAllocationStrategy>("signed_distance_allocator", node_allocator)};
+      s_allocator_id = signed_distance_allocator.getId();
+    }
+
     rc = internal::read_stl_mesh_shared(file,
                                         comm,
                                         s_allocator_id,
@@ -161,7 +162,6 @@ int signed_distance_init(const std::string& file, MPI_Comm comm)
   {
     rc = internal::read_stl_mesh(file, s_surface_mesh, comm);
   }
-
 #else
 
   SLIC_WARNING_IF(Parameters.use_shared_memory,